✨ hex formatting in uart_println

+++ bootloader/include/string.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 int strcmp(const char *, const char *);
 
-char* itoa(int val, int base);
+char* itoa(int64_t val, int base);
